Fix FoldPHIEntries merge error (#4723)
The FoldPHIEntries function is a generalization of FoldTwoEntryPHINode
in the upstream code. As such, changes to the upstream code are merged
into the Intel-customized version.

Part of the generalization in the Intel code is to loop over BB
predecessors looking for folding opportunities where the upstream code
only makes one pass. There was a change in the upstream code that would
make an early exit under some circumstances and not perform any
transformation. In the Intel code, we may have already performed some
transformation before this condition is reached, so it is not
appropriate to unconditionally return 'false' at that point as the
upstream code does. Incorrectly reporting that no changes to the CFG had
been made was causing a crash when using the new pass manager because
the new pass manager (in prod and debug builds) verifies that no changes
were made.

This patch replaces the early return with a 'continue' which will look
for further opportunities and preserves the state of the 'Changed'
variable.