[NFC] CMPLRLLVM-37183: Replace MDNode::concatenate with use of ArrayRef (#4721)
* CMPLRLLVM-37183: Replace MDNode::concatenate with

As CJ noted in CMPLRLLVM-36389,MDNode::concatenate needs 2 MDNodes, so a
temporary node is created. MDNode::concatenate creates a new MDNode from the
operands of the 2 nodes. The temporary node is leaked. There is no easy way to
delete the temporary node, as there is no public interface to delete MDNodes.
Best way to fix this, is to create the new node directly with an ArrayRef
of operands, instead of using MDNode::concatenate.

I have based this fix for InlineFunction.cpp on CJ's recommendation above.

* Fix Copyright