[OpenCL] Check the existence of runtime config file (#4461)
* [OpenCL] Check the existence of runtime config file

We suppose there must be a config file which contains CL_CONFIG_DEVICES information.

It's hard to check whether there is a config file as it's loaded in on constructor.
And the constructor is used during global variable initialization. We cannot use
exception to handle it. So we check the CL_CONFIG_DEVICES information in clGetPlatformIDs.

* [OpenCL] Move the config file and device type check to module implementation

* [OpenCL] Move the check to module initialization phase.