[NFC][VPlan][HIR] Add test to reveal a bug that produces broken SSA after HIR decomposer's fixPhiNodes. (#4740)
We rely on DDG to place initial PHI nodes for ambiguous symbases
and this can lead to broken SSA when the edges in DDG are
inaccurate. This patch adds a test that shows this bug where a
redundant PHI that blends its own value along on of the edges
is generated. The bug will be resolved in follow-up PR.

JIRA: CMPLRLLVM-36880
Change-Id: I4a2a41218417dc91e1a75341b9c4829ef4959b8d