[DTRANS] CMPLRLLVM-36026: Fix for bad pointer manipulation for related types after offset-based argument promotion (#4745)
* [DTRANS] CMPLRLLVM-36026: Fix for bad pointer manipulation for related types after offset-based argument promotion

The argument promotion pass can promote access to a structure's fields.
The issue is that this pass introduces byte flattened GEPs. These
GEPs are used to compute the field that is being accessed, or the
fields of the inner structures.

The PR #4625 fixed a bug in the argument promotion, which allowed to
run the pass in one function for 541.leela, but it broke the DTrans
analysis for related types. The issue is that the pointer operand of
byte flattened GEP won't have a dominant aggregate type, but we know
that it is used with related types and we can collect the encapsulating
structure. This change set adds a function that verifies if the
structure was collected cortrectly and if the offset of GEP covers
all the fields accessed.

* Fix for Robert's comments

* Fix for Chris' comment