[DTrans - OpaquePtr] Rework the PtrTypeAnalyzer interaction between analyze and infer behavior (#4760)
* [DTrans - OpaquePtr] Rework the PtrTypeAnalyzer interaction between analyze and infer behavior

[CMPLRLLVM-36511]
This is a fix for some bad compile time issues that could occur with the DTrans
PtrTypeAnalyzer caused by a value that needs to be inferred by looking all the
users of the value, which would then trigger analysis of values that could cause
more infer calls, and so on, which led to the same value being processed many
times because it was a common dependency in the chain. The change redesigns this
interaction by no longer calling infer functions from the analyze functions, and
no longer calling analyze functions from the infer functions. Also so some code
that was previously in place to handle this messy interaction to prevent infinite
cycles is removed. In this revised implementation, the dependency stack is extended
to hold both the analysis dependencies and the infer dependencies. And the
construction of this stack is updated to populate the stack for the infer dependencies
This is mostly an NFC change because it is not changing the analysis result, but
is changing the way it gets to the result, so most of the existing test cases
cover this change.

One fix that is made is to detect a case where a value uses a load/store of
the pointer location of the value itself. In this case, because it is expected
that the pointer operand is a pointer type of the value operand, this would lead
to additional levels of pointer level indirection being tracked for the value,
because an additional pointer dereference level would be added each time the value
was visited.

* Fix for Robert's comments. Also, remove some #if 0 code that was no longer relevant.

* Fix for Prasad's comments