[NFC] Use LLVMContext interface function to check for opaque pointers for DTrans base class (#4806)
Update the new DTrans code to just use the setting in the LLVMContext
to identify if typed pointers or opaque pointers are being used. When
the code was being developed, the community had not implemented a
way to check for typed vs opaque pointers, so the PtrTypeAnalyzer
was tracking this. This is no longer necessary, so this change is
eliminating the unneeded code, and letting the DTrans base class
just check the context value.