[DTRANS] CMPLRLLVM-36026: Extend arrays with constant entries to collect boost arrays, and relax some conditions to collect the data (#4850)
* [DTRANS] CMPLRLLVM-36026: Extend arrays with constant entries to collect boost arrays, and relax some conditions to collect the data

This change set extends the arrays with constant entries to handle
the case where the field is a boost array. Arrays made with boost
libraries are a structure that have one field and that field is
an array. We are going to find the case where the GEP that
access this field is followed by a GEP that access the actual array.

Also, this change set relaxes some conditions to collect the
arrays with constant entries.

* Fix for Chris and Robert' comments