[DTRANS] CMPLRLLVM-37427: Fix for possible segfault when collecting arrays with constant entries (#4902)
* [DTRANS] CMPLRLLVM-37427: Fix for possible segfault when collecting arrays with constant entries

The PR #4850 introduced a possible segfault. There is a chance that
a GEP doesn't have a user but the analysis process tries to access
GEP->user_back(). This produces a segfault.

This change set makes sure that there are users before accessing
the information, as well it traverse through all possible users
of the GEP. Also, It fixes a possiblke segfault when the DTrans
immutable pass is collecting information. There is a chance that
the entries collected in the array are larger numbers than the
map's size. We are going to just copy the entries that aren't null
and then call std::sort to simplify the implementation.

* Fix for Robert's comment

* Fix for Robert's comment