[LLD-ELF] CMPLRLLVM-37502: Identify ELF target for GNU LTO objects (#4939)
* [LLD-ELF] CMPLRLLVM-37502: Identify ELF target for GNU LTO objects

The community recently moved the identification of the ELF object
when calling invokeELFT. The function that sets the ELF target runs
after identifying the input objects. This produced an llvm_unreachable
assert when calling invokeELFT without -m and GNU Lto objects since the
ELF target is not set. This change set checks if -m was specified, then
use the configuration from -m, else goes through the GNU LTO input
objects and identify the ELF target used in the input files to
pass the right ELF target to doGnuLTOLinking. Guilty commit:

https://reviews.llvm.org/D118551