[LLVM][Casting.h] Don't create a temporary while casting. (#5003)
C-style casting can create a temporary when compiled by a C++ compiler, which was emitting a warning casting a reference to another reference. We can't use C++-style casting directly because it doesn't always work with incomplete types. In order to support the current use-cases, for references we switch to pointer space to perform the cast.

https://reviews.llvm.org/D125482