[CMPLRLLVM-37304] Clone opt-report metadata during function cloning (#4971)
Optimization report metadata is dropped during basic block cloning to
avoid duplication of optimization report output, but when cloning
functions or modules it would be better to clone this metadata into the
new funciton/module instead so that it is also visible there. Besides
making optimization reports for cloned functions more complete, this
also avoids issues in tools like llvm-reduce which assume that when a
module is cloned it will have the same amount of metadata.