[VPlan][HIR] Use getNestingLevelFromInsertPoint() interface in more makeConsistent calls (#4893)
* [VPlan][HIR] Use getNestingLevelFromInsertPoint() interface in more makeConsistent calls

As suggested in review comments for PR#4679, the changes here start using the
new interface added to get insertion point nesting level when making a DDREf
consistent in all places where this appeared like the right thing to do. This
is shown to address a potential issue in the included LIT test.

* Changes to fix the failure seen in alloy testing.

When calling the widenRef which takes a RegDDRef from widenRef interface
which takes a VPValue, the RegDDRef passed is uniform. We now pass this
information to the callee so that we do not unnecessarily try to invoke
codegen for nested blobs. In addition to avoiding unnecessary instructions,
this also addresses issues such as UMIN_SEQ scev type nested blobs which
can be seen in VPValue based CG and which are not supported in NestedBlobCG
which was meant to be used only for mixed mode code generation.