[DTrans - OpaquePointers] Improve PtrTypeAnalyzer collection for Global Variables with elided GEPs (#5005)
Following the change to perform constant folding on GEP.x.0.0 to eliminate GEP
operators, analysis of the effective types is complicated when a PHI or select
node input is a global variable that is an aggregate type. The resolved type for
the global could represent the type of the global variable or element zero
within it. Adding both types as the resolved type can result in safety flags
or uses of the merged variable not being able to be analyzed. Instead, we will
use the context of the other values in the SelectInst/PHINode to resolve
whether to use the type of the global variable, or the element zero type.

This is one of the changes required for CMPLRLLVM-36882