CMPLRLLVM-37390: Suppress outer big 'branchy' IVDep loops (#4986)
This is a workaround to prevent compilation hang in HIROptPredicate.
The pass is trying to hoist many invariant conditions out of a huge
outermost IVDep loop. The hang is in the logic to update top sort
number of nodes during insertion. Ideally, we should improve that
logic but the change is risky for product branch hence the
workaround.