[VPlan] Changes to fix crash in vector code generation when broadcasting pointer returned from a call that can throw (#5075)
When generating code to see if dynamic peeling is needed, vector CG does
a broadcast of the pointer value at the point where the value is defined.
In cases where the pointer is the return value from a call that can throw,
we were trying to generate the broadcast after the corresponding invoke
instruction which needs to be the last instruction in the block. The fix
generates the broadcast in the pre-header.