CMPLRLLVM-37668: Test for null clones before performing multiversioning (#5084)
* CMPLRLLVM-37668: Test for null clones before performing multiversioning

525.x264 without full base options does not use DTrans to resolve
indirect calls to direct calls. Consequently, while there are calls to
mc_chroma at the time call tree cloning is applied, none of them is
actually cloned in call tree cloning. The early return test for
MultiVersionImpl::doCodeGen() needs to run before the body of the original
function is deleted, or the original function will be converted into a
declaration, even though no cloning is done.

* Fix blank lines

* Fix blank line