[LLD] Enable vectorization for LTO under new pass manager (#5070)
As reported in https://jira.devtools.intel.com/browse/CMPLRLLVM-37498,
there are some serious performance regressions on Windows with the new
pass manager during LTO.  The problem is in LLD, which is responsible
for setting up the LTO configuration.  Vectorization was enabled at some
point for the ELF path, but not for COFF or MachO.  Enabling this along
the COFF path is needed for Windows.

The differences between ELF, COFF, and MachO should be brought to the
attention of the community maintainers, as discussed in the tracker.
For now, fix this specific issue as an Intel customization.