[DTrans - OpaquePointers] Extend PtrTypeAnalyzer for GEPs that index into arrays of pointers that are element-zero members (#5085)
A GEP instruction may use the element-zero type as the source type when
the pointer type is a pointer to a structure. When this is done to index
into an array of pointers, the final operand of the GEP may be non-constant.
The PtrTypeAnalyzer was treating this as an unhandled case. This extends the
analysis in the PtrTypeAnalyzer to handle some cases where an appropriate
type can be identifed by traversing element zero aggregate types.

This is one of the changes required for CMPLRLLVM-36882