[VPlan][HIRCG] Fix to correctly mark livein temps assigned to uniform IF (#4889)
The map used to keep DDRefs assigned to VPExternalDefs is extended so we can keep
DDRefs for uses of VPExternalDef in several different loops.

The VPValue->Value maps used in CG contain only one copy of VPValue. That
is true for VPExternalDefs as well. Usually we have one symbase for
all copies of VPExtrenalDef but in case if uniform IF we generate code
several times (when have peel/remainder) and have different symbases
assigned each time. Marking one temp as livein in all loops works
correctly when in all uses the VPExternalDef has the same symbase.
It does not work in case of uniform IF. The fix is to add a special map
for VPExternalDefs and put there all the temps created for them during CG.