[VPlan][Scan] Bail out from scan by default (#5136)
After FE added support for openmp simd scan reduction pragmas
the vectorizer started to crash as scan support is not implemented in
VPlan yet. This patch implements a bail out, and provides a developer
toggle to override it. Toggle default value is intended to be flipped
once the feature is mature.

The test exercises only the IR path, as at this point HIR framework
bails out both due to unrecognized scan pragmas and having a fence
instruction in the incoming IR by current design.