[DTrans - OpaquePointers] Enable field modref to return results to loop analysis for 538 (#5174)
* [DTrans - OpaquePointers] Enable field modref to return results to loop analysis for 538

This finalizes the reporting of the NoModRef result to the loop optimizer for 538 with the
current IR under opaque pointers. Pointer type analyzer is updated to handle i8* type
returned from allocation call when analyzing a GEP that then access the element zero
array element. The DTrans safety analyzer special case is updated to handle the uses
of the allocated pointer for the 'double*' field in the structure.

[CMPLRLLVM-36882]

* Fix comment