[DVCP] CMPLRLLVM-37355: Perform DVCP for dope vectors that are local to a function (#5189)
* [DVCP] CMPLRLLVM-37355: Perform DVCP for dope vectors that are local to a function

This chenge set extends the DVCP to perform the optimization on
dope vectors that are local to a function and won't be used outside
of the function, neither passed to another function. The changes
reuses the analysis and propagation processes that were already
implemented. The analysis process was relaxed to handle the case
when an instruction is a BitCast and is used to initialize the
pointer as null.

* Fix for Chris' comments

* Fix for Robert's comments