[DTrans - OpaquePointers] Fix PtrTypeAnalyzer dominant type computation for pointer-to-pointer handling of element zero. (#5324)
* [DTrans - OpaquePointers] Fix PtrTypeAnalyzer dominant type computation for pointer-to-pointer handling of element zero.

This is a fix to the PtrTypeAnalyzer dominant type computation for pointer-to-pointer types
that are also used as a pointer-to-pointer to the element zero type. Previously, the code
that checked for this would still report the type as ambiguous because it did not set the
flag used for safe element zero types. Also, fix the code for this to be deterministic
regardless of whether the pointer-to-pointer type is seen as the container type or the
pointer-to-pointer to element zero type first.

[CMPLRLLVM-37908]

* Make pointers opaque

* Fix comment typo