[SLP][NFCI] Remove Intel customization in tryToVectorizeHorReductionOrInstOperands. (#5217)
* Remove unnecessary (and buggy) Intel customization.
The case is now already handled by the community vectorizer in a
slightly different manner (meaning we don't need to customize).
Additionally, was the source of regression.

* Re-enable horizontal-list.ll and adjust cost customization to be high enough.
Reverting broken customization in tryToVectorizeHorReductionOrInstOperands()
not only fixed a bug there but also exposed bug in another customization
that was put in with original PSLP implementation (the way we bypass vectorizing
tiny trees). The customization at the moment looks redundant as well and is planned
for removal. Unless that happened increase vectorization prohibitive cost value to
work around the issue with the customization.

JIRA: CMPLRLLVM-37730