[DTrans] Make DTrans use typed passes or opaque pointer passes based on input IR (#5356)
* [DTrans] Make DTrans use typed passes or opaque pointer passes based on input IR

This is one step along the way of elimninating the need for special flags for
DTrans when the compiler is using opaque pointers.

The change enables the DTrans transformations to b based on the input IR for whether
to use that typed pointer and opaque pointers pass. Both passes will be scheduled
to run, but one of them will be a no-op. This is to prepare for migrating to just
having the opaque pointer passes. This removes the flag
-dtrans-force-opaque-pointer-pipeline. It introduces the flag
-dtransop-allow-typed-pointers to force the opaque pointer passes to be run on
non-opaque pointer IR for testing purposes. LIT tests are updated to add this
flag for the cases where the tests were created before opaque pointers were
available, and have not been migrated to only running in opaque pointer mode.

* Fix for Robert's comments

* Fixes for Axel's comments

* Fix for Axel's comments