[NFC] CMPLRLLVM-37998: Fix use of isa and dyn_cast in IPO Prefetch for Klocwork (#5373)
This change set is a code cleaning for the function IsLocalStore in the
IPO Prefetch. The original form use a call to isa<> and then dyn_cast<>.
This is redundant because if isa<> passes, the dyn_cast<> should pass.
Yet, Klocwork triggered in the call to dyn_cast with a possible
null return.