CMPLRLLVM-37425: Prevent infinite recursion when the attributor tries to simplify a value (#5426)
* CMPLRLLVM-37425: Prevent infinite recursion when the attributor tries to simplify a value

The attributor pass have two functions that are mutually recursive:
reproduceValue and reproduceInstruction. The issue is that they
can get stuck in an infinite recursion due to PHI nodes that
use the values from each other. This change set identifies if a
Value was already visited, then return it since it won't be
able to find any more information.

Also, it fixes a possible issue where the SSA form can be broken
after applying the simplifcation.

* Add test case for opaque pointers.