[CMPLRLLVM-37927] Fix ZTT crash in HIRRowWiseMV (#5355)
* [CMPLRLLVM-37927] Fix ZTT crash in HIRRowWiseMV

addZttPredicate only adds a predicate if there's already a ZTT; if
there's not already a ZTT, it asserts. For HIRRowWiseMV, this led to a
crash if the inner loop does not have a ZTT but the outer loops do. I've
fixed this by adding outer ZTTs to an enclosing HLIf instead.