CMPLRLLVM-37323 "510.parest compfail when using -g and opaque pointers" (#5448)
* CMPLRLLVM-37323 "510.parest compfail when using -g and opaque pointers"

In LLVM, types are part of the scope hierarchy and need to be traversed to
locate the containing local/global scope. Not doing so can cause type
information to be duplicated unnecessarily. In very rare cases, the duplicate
type information can include copies of member function arguments which lead
to a crash during dwarf emission.

* Clean up the metadata numbering in the LIT test.

* Fix spelling error "Exraction" --> "Extraction".