[OpenCL][NewPM] Drop TBAA metadata from OpenCL SPIR 1.2 bitcode (#5416)
OpenCL CTS SPIR 1.2 bitcode is compiled with LLVM 3.2 and its TBAA
metadata doesn't have constant default value in the third field,
unlike current LLVM. The two versions of metadata could be mixed after
BuiltinImport pass. This causes correctness issue as types are
considered as no-alias, so we drop TBAA metadata in this situation.