[OpaquePointers] Fix argument promotion on callback functions when using opaque pointers (#5456)
The change in https://github.com/intel-restricted/applications.compilers.llvm-project/pull/4271 was preventing argument
promotion on callback functions because the argument type mismatch check used the argument type of the argument number
at the callsite, without adjusting for the possibility that the callsite was an abstract callsite. This inhibited the
argument promotion on the callback functions for OpenMP when using opaque pointers. This change removes the special
code that was added to xmain, because the community code resolved the problem of signature mismatches (and also handled
case where the return value type mismatches) in https://github.com/llvm/llvm-project/commit/51561b5e8017a3153629ba45b89d013ffa665f6c.

The test case added with PR #4271 is left in to ensure testing of the original case that was found is still covered.

[CMPLRLLVM-38139]