review/skozhukh/blend safe val (#5292)
* [VPlan][HIRCG] Serialize integer div/rem when divisor is not safe.

Currently we bailout when we see integer div/rem under an explicit
condition. But the bailout code does not account masked mode loop
which adds an implict mask. Implementing serialization to avoid
masked mode rejection.

* [VPlan]Implementing blend-with-safe-value transformation on masked integer div/rem.

The transformation is to fill-in masked lanes of divisor with a
safe value (which is 1) when the divisor is unsafe. I.e. when
divisor is not a constant or is an unsafe constant (0 or, for sdiv, -1).
The transformation allows generating the division unmasked avoiding
unexpected exceptions coming from masked lanes.

* Responding review comments

* Responding more review comments.

* Responding review comments