CMPLRLLVM-5589: Negative test for recognition of user alloc function which (#5472)
does not free the memory pointed to by the argument of the user call.

This is similar to the LIT test ptrtype-useralloc01neg.ll, but uses a
GEP with a structure type. We will not fix the handling in the original
DTransAnalysis because, as Chris noted, it will become obsolete when we
switch to opaque pointers.