CMPLRLLVM-264: Add support for trace to indicate which types are C-rule compatible. (#5476)
* CMPLRLLVM-264: Add support for trace to indicate which types are C-rule
compatible.

CMPLRLLVM-264 includes an example where it was thought that two types were
incorrectly discovered to be C-rule compatible. In our current compiler,
this error does not exist. However, I thought it would be useful to add a
trace and some LIT tests to more easily check if DTrans thinks two types
are C-rule compatible.

* Fix typos

* Fix another typo in comment

* Be consistent on use of ptr