[OpenCL][LIT] Temporary disable test cases (#5509)
Same as #4489, We temporarily disable these case because unexpected failure
when test framework is migrated to an older kernel. We will enable it after
analysis.

The two test backend test cases is already disabled before. I don't know why
it's reported again.  I disabled it for for all platform this time.