[DTrans - OpaquePtr] MemManageTrans matching of 'StringAllocator' structure (#5513)
This is the first change in converting the code for the MemManageTrans pass to
support opaque pointers.

This change performs matching of the 'StringAllocator' structure which is needed
to mark the inlining preference on some functions during the pre-LTO compilation
phase. The code is basically the same as the original MemManageTrans code seen
in MemManageInfoInfo.h, but modified for the functions to work based on examining
the structure types based on the DTrans type metadata. One test case is just the
original LIT test for with updates to insert the DTrans type metadata. The other
LIT test is created based on the IR that is currently seen when compiling with
opaque pointers.

[CMPLRLLVM-36022]