CMPLRLLVM-37752: Disable DVCP for lower bound field in global dope vectors (#5502)
* CMPLRLLVM-37752: Disable DVCP for lower bound field in global dope vectors

This change set disables the const prop for the lower bound fiel
in global dope vectors if there is at least one function with 7
or more call instructions that contain the attribute
"prefer-inline-tile-choice". The reason is that appliying DVCP
in this case complicates the analysis process for loopopt. Also
this change set turn on by default aggressive DVCP.

Gains:
  polyheadron2011F/fatigue2 -> +31.3%
  cpu2017/649.fotonik -> +2.2%
  specACCEL/551.ppalm -> +4.6%
  specACCEL/563.pswim -> +3.0%

Known regressions:
  specACCEL/556.psp -> -2.7%
  spec_omp2012/371.applu331 -> -6.7%

* Fix for Chris' suggestion

* Fix for Chris' suggestions