[SLP] Account for cost of lost scalar FMAs when vectorizing (#5454)
A previous patch accounted for the cost of vectorizing a horizontal
reduction of adds fed by a vectorizable tree of multiplies.  There is
also the more general case of a vectorizable tree of multiplies that
each feeds an independent add.  This patch consolidates the costing
logic so it can be used for both cases.  Per community suggestion, as
much of the cost calculation as possible is moved into TTI.  An
additional test variant is added, and the test case is renamed
appropriately.