[Driver] Sanitizer flags not compatible with OpenMP target (#5515)
When using -fsanitize with the spir64 target for OpenMP, we emit an
error diagnostic.  This prevents any kind of -fsanitize usage with
the presence of offloading.

Update the behavior to ignore the sanitize flags for the target compilation
if the target is spir64 based.  Also emit a diagnostic alerting the user
that the flag is not used for the target compile.