CMPLRLLVM-38271: Disable as multiversioning target if the virtual call function type doesn't matches the target function type (#5611)
* CMPLRLLVM-38271: Disable as multiversioning target if the virtual call function type doesn't matches the target function type

This change set adds an extra check when applying devirtualization. It
checks if the funtion type of the virtual call mismatches the function
type of the target function. If not then don't add the target but include
the default call. Also, it updates the documentation and the function
that identifies if a function have direct calls in order to apply
partial inlining.

* Fix for Robert's comments

* Fix for Robert's comments

* Fix for Robert and Prasad comments