[VPlan][CM][NFC] Change usage model for Invalid/Unknown Instriction Costs. (#5442)
* [VPlan][CM][NFC] Change usage model for Invalid/Unknown Instruction Costs.

With this change both Invalid and Unknown costs become poisonous
and propagate through arithmetic. However Unknown cost is filtered
out during CM on VPInstruction level and does not cause the whole
VPlan to be of Unknown cost. Unlike Ivalid cost, which propagates to
Vplan level.

Unknown Instruction Cost is reserved for case when vectorization is
valid but given instruction cost can not be evaluated for some
reasons such as unknown costs of elements of a compound
instruction or just yet missed cost modelling.

Invalid cost is reserved for overflow problem only and does not block
vectorization if it is forced. Otherwise Invalid cost Plan are not chosen
for vectorization though (with exception when every available Plan has
Invalid cost).