38460: Allow LCSSA phis to live past SCEVExpander, as they can still be deleted by IC. (#5698)
* 38460: Allow LCSSA phis to live past SCEVExpander, as they can still be deleted by IC.

* Wrap code into #ifndef NDEBUG

* Update intel-scexp-lcssa-fix.ll

* Declare InsertedSet permanently

The original fix, prevents LCSSA phis from being deleted, when there are uses that will not be deleted. This skips the assertion, and hopefully prevents incorrect code from being generated also. This "protection" must be run all the time.
I suspect that all the uses in the LCSSA web are actually dead, so we could delete the phi and leave an undef value, but I can't prove that right now.

Co-authored-by: Haohai Wen <haohai.wen@intel.com>