[OpenMP] Fix firstprivate handling on combined distribute directives (#5724)
For a combined construct such as 'target teams distribute' we generate
multiple directives. The general rule is to apply the clause to each
directive that allows that clause. Since 'target', 'teams', and
'distribute' all allow 'firstprivate' it was added to all three.

There are special rules in the OpenMP spec (5.2/sec.17.2) that need to
apply instead of the general rule.  In this case 'firstprivate'
clause is applied to 'distribute' but not 'teams'.

A new routine has been added (isAllowedClauseForDirectiveFull) to
handle cases like this where the full construct needs to be
considered.

Handling of more special cases can be added in the future, but this
change only handles the firstprivate/teams case to avoid
destabilizing the compiler.

Fixes: CMPLRLLVM-37965