[CMPLRLLVM-30767] Attempt cross-NUW/NSW improvement in getSign/ZeroExtendExpr (#5775)
As an optimization, getZeroExtendExpr only attempts to prove NUW on
AddRecExprs, and getSignExtendExpr only attempts to prove NSW.

However, xmain improvements may allow proveNoWrapViaConstantRanges to
prove, e.g., NSW when NUW is already known. In CMPLRLLVM-30767, the
community introduced an improvement which proved an additional NUW
elsewhere in SCEV analysis. This resulted in getZeroExtendExpr skipping
proveNoWrapViaConstantRanges and the loss of an xmain-specific NSW.

Fixing CMPLRLLVM-30767 only requires the modification to
getZeroExtendExpr, but for symmetry we make the equivalent change in
getSignExtendExpr too.