[OpenMP] Adjust in_reduction code after community implementation (#5856)
Remove the code that gives an error if in_reduction is used with
target. This was added when we implemented late-outlining and
community did not have it yet.

Also disable the code adding .task_red variables. These are not used
with late-outlining but if they are captured it interacts with the
map code causing asserts.

Fixes: CMPLRLLVM-38687