38725-1: Special handling for phi nodes in the new createSCEVIter routine. (#5925)
When SCEV was changed to be partially iterative, the return value of createSCEVIter is taken from the cache. PHI node SCEVs are represented in the cache with a SCEVUnknown instead of a proper AddRec. We need to return the real SCEV instead of using the cache. This matches what the recursive createSCEV is doing. Fixes intel-identical-phi-operand-scevs.ll.