[DVCP] CMPLRLLVM-38321: Disable aggressive DVCP if prefer-inline-tile-choice is available (#5919)
* [DVCP] CMPLRLLVM-38321: Disable aggressive DVCP if prefer-inline-tile-choice is available

This patch reverts #5502 (disable lower bound dvcp) since it produces
a regression in alder-lake systems. Rather than disable the lower bound,
we are going to identify if prefer-inline-tile-choice is available. If
so, the disable aggressive DVCP. The reason is that there could be
a dope vector that is NOT related to the specialized tiling, but
it lives in a function that is selected for prefer-inline-tile-choice.
If we disable DVCP for these dope vectors then we loose some vectorization.

Gain: +2% - +3% CPU2017/549.fotonik in alder-lake systems

* Fix for Robert's comments