[DTrans - OpaquePtr] Inhibit the DTrans safety analyzer when the DTrans type metadata is invalid (#6041)
This extends the checks for invaid DTrans metadata to include the case where there are
conflicting type descriptions for a structure type. In this case, the type for a field
would not be uniquely identified. This should not arise in general, however problems
with the IRMover merging of types could cause it resulting in stability issues with the
DTrans safety analyzer.

[CMPLRLLVM-38859]