[VPlan] Do not drop FMF flags for in-memory reduction finalization (#6006)
The FMF flags are lost for in-memory reductions on the finalizing
horizontal instrinsic. This is not correct for fp consistency.
For inscan reduction this is a potential performance problem:
as the scan computation is inside the loop and inscan reduction is
always in-memory.

Since we do not have any linked values at the moment of checking,
we need to analyze the values stored into the reduction alloca.