[DTrans - OpaquePtr] PtrTypeAnalyzer analysis of flattened GEPs (#6111)
* [DTrans - OpaquePtr] PtrTypeAnalyzer analysis of flattened GEPs

This extends the GEPs analysis within the PtrTypeAnalyzer to try to
resolve addressing of structure fields when the indexing is done
using a GEP with a single-value type (ptr, i32, double, etc), rather
than normal structure indexing. This handles the case where the
indexing amount is a constant that is smaller than the size of the
structure, and a field can be identified as being at that address.
This typically is seen when the type of the field being accessed
matches the indeing type. For now, the safety analyzer will mark
these structures as 'unhandled' because this does not occur on the
types that we need transformations for.

[CMPLRLLVM-38875]

* code cleanup

* Update comments

* Update comments