[CMPLRLLVM-38803] llvm-extract to set DTrans type metadata when converting 'alias' to 'declare' (#6157)
* [CMPLRLLVM-38803] llvm-extract to set DTrans type metadata when converting 'alias' to 'declare'

Extend the support for preserving DTrans type metadata when using llvm-extract to extract
functions for test cases. Previously support was done when a 'define' is converted to a
'declare'. This change handles the case where an 'alias' is converted to a 'declare' because
an 'alias' is often used when C++ generates multiple constructor funtions, and one is
just created as an alias to another.

* Remove replaced code

* Update based on review comments