[NFC] [DTrans] Correct logic in MemManageTransImpl::recognizeConstructor (#6141)
This fixes the logic in the function MemManageTransImpl::recognizeConstructor
for when the counters are incremented when identifying the assignments to
the List.Head and List.FreeHead fields of the ArenaAllocator. It is NFC,
becuase the expected behavior is that each will be incremented once, and
this just fixes the logic to make the increment happen agree with the match
of the pattern.