[AndersensAA] (CMPLRLLVM-38623) Propagate UniversalSet to points-to sets of all dereference levels of a pointer If UniversalSet is in point-to set of the pointer (#6132)
* Propagate UniverselSet to points-to sets of all dereference levels of a pointer If UniverselSet is in point-to set of the poinrter

While solving constraints, if UniversalSet is in points-to set of CurrNode,
it propagates UniverserSet to the other deference levels of the pointer (if
it is not already there).

Fixed some LIT tests that fail due to this change.

This causes performance regression:
EEMBC regression: Need to handle user routine t_printf as wrapper for vsprintf so that arguments of t_printf calls are not treated as conservative.

spec_omp2012ref/371: Handled AbstractCalls to improve points-to but still there
is perf regression (8%). Can't reproduce the regression on my local machine
using zperf. Will look into it later.

* Fix comments made by Robert and Chris.