CMPLRLLVM-35880: Inline "relax" for NavigationDelta in Geekbench6 Version 31 (#6156)
* CMPLRLLVM-35880: Inline "relax" for NavigationDelta in Geekbench6 Version 31

This limits the size of single ODR functions that can be inlined so that "relax"
gets inlined. Different limits are used for Windows and Linux because the
implementations of the STL are quite different.

This also includes a bug fix for the metadata inlining report. I encountered
the bug while writing the LIT test. It was assumed that all dead functions
have local linkage and can therefore be marked 'L', but there are different
types of local linkage: for example, static and link once ODR. So we must save
the linkages in the function metadata so they can be used when annotating the
callsites during inline report emission.

* Fix for Chris's comments

* Fix comment

* Add PrepareForLTO to guard against regression